Fixed Level 7 Issues and scoring
Scoring now works correctly and has a max of 1000. It is 75% based on number of attempts and 25% based on time to complete. Level 7 also now works correctly by creating a line of goals that are offset from the origin. Further testing should be done to determine if it is still possible to solve with the offset. Finding an acceptable graph within bounds is tricky for levels 6 and 7, especially on higher difficulties, so the player must be patient (in my testing it has always eventually found an acceptable solution). It may be worth finding a way to only show a loading screen while the program tries different combinations of vectors as the user may think that the program is glitching  with the game as is (it reloads the scene each time a goal is forced outside of bounds).